dip_License_plate_detection/requirements.txt
tk==8.6.10
Pillow==8.3.2
opencv-python==4.5.3.56
numpy==1.21.2
pytesseract==0.3.8
matplotlib==3.4.3
